requirements.txt
clickhouse_driver==0.1.3
pulsar_client==2.8.0
kudu_python==1.2.0
fastavro==1.4.5
typing~=3.7.4.1
pymongo~=3.9.0
impyla~=0.16.2
retrying~=1.3.3